requirements.txt
absl-py==2.2.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
Authlib==1.4.1
blinker==1.9.0
boto3==1.36.26
botocore==1.36.26
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.1
Deprecated==1.2.18
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
envs==1.4
flask>=3.1.1
jinja2>=3.1.6
Flask-Limiter==3.12
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Sitemap==0.4.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
future==0.18.3
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-core==2.25.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.3
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.73.1
gspread==6.1.4
gunicorn==23.0.0
h11>=0.16.0,<0.17
httplib2==0.22.0
httpx==0.28.1
idna==3.10
immutabledict==4.2.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
limits==5.0.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.4
oauth2client==4.1.3
oauthlib==3.2.2
ordered-set==4.1.0
ortools==9.12.4544
packaging==24.2
pandas==2.2.3
praw==7.8.1
prawcore==2.4.0
protobuf==5.29.5
psycopg2==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycognito==2024.5.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose[cryptography]>=3.3.0
pytz==2025.1
redis==5.2.1
requests>=2.32.4
urllib3>=2.5.0
httpcore>=0.20.0
requests-oauthlib==2.0.0
rich==13.9.4
rsa==4.9
s3transfer==0.11.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
update-checker==0.18.0
upstash-redis==1.4.0
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
WTForms==3.2.1
zipp==3.23.0
